Add files via upload
Sentiment analysis has been a topic of many researches and interest. In this article, computational steps are shown and worked out to show, how we can have multiple centres in the sentiment classification task. This does not stop the existing works but provides more avenues to open research applications more. Conceptualize a 5-class classification task for sentiments, what it means to look in more than three-dimensional spaces, and what it means to have more insights. This can be taken as an ML exercise, both practical & theoretical, till it’s real applications are out. However, major application is that new words can be assigned sentiment without a supervised approach, but only via Trained Large Language models. The 3-class sentiment orientation is well defined as till now we consider subjectivity in three spheres. But consider the case, of very high positivity, positivity, neutrality, and very negative to negative sentiments. Hence the need to define, more than three typical classes. Here we provide the algorithm for assigning the multi-class sentiments and their degree for any new word so that not much of resources are used.